Undo ref hub, it leads to issues with std/isolate
Explicitly in the example
ex_stdinput_no_server.nim

Spawning tasks requires isolation.
However, tasks that want to call back once they're done need the channelhub,
which would lead to 2 threads sharing the same ref to an instance of channelhub, which can't be isolated.
If the channelhub is a value object then it just can get copied around willy nilly.

This is fine, because I demand explicit destruction of it via a destroy proc anyway at the end of the program.